#199 - refactors query-param.ts naming
